更新calcite分支

+++ common/src/main/java/io/mycat/beans/resultset/MycatUpdateResponseImpl.java
@@ -1,3 +1,17 @@
+/**
+ * Copyright (C) <2020>  <chen junwen>
+ *
+ * This program is free software: you can redistribute it and/or modify it under the terms of the
+ * GNU General Public License as published by the Free Software Foundation, either version 3 of the
+ * License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
+ * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with this program.  If
+ * not, see <http://www.gnu.org/licenses/>.
+ */
 package io.mycat.beans.resultset;
 
 import java.io.IOException;
@@ -5,10 +19,10 @@ import java.io.IOException;
 public class MycatUpdateResponseImpl implements MycatUpdateResponse {
 
   final int updateCount;
-  final int lastInsertId;
+  final long lastInsertId;
   final int serverstatus;
 
-  public MycatUpdateResponseImpl(int updateCount, int lastInsertId,int serverstatus) {
+  public MycatUpdateResponseImpl(int updateCount, long lastInsertId,int serverstatus) {
     this.updateCount = updateCount;
     this.lastInsertId = lastInsertId;
     this.serverstatus = serverstatus;
